+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Implemented `std::convert::[Try]From` for `NonMax*` & primitive types to match `NonZero*`
 * Implemented `std::fmt::{Display, Binary, Octal, LowerHex, UpperHex}` to match `NonZero*`
 * Removed outer "NonMax(...)" in output from `impl Debug for NonMax*`, matching `NonZero*` behavior
+* Implemented `std::str::FromStr` for `NonMax*` to match `NonZero*`
 
 ## 0.4.0 (2020-09-27)
 * Raised MSRV to 1.46.0 to make more methods `const`.